cvbuilder-backend/requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
Django==4.2.16
djangorestframework==3.15.2

# ─── Database ─────────────────────────────────────────────────────────────────
mysqlclient==2.2.4

# ─── Authentication ───────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ─── Environment & Configuration ──────────────────────────────────────────────
python-decouple==3.8

# ─── CORS ─────────────────────────────────────────────────────────────────────
django-cors-headers==4.4.0

# ─── Audit Logging ────────────────────────────────────────────────────────────
django-auditlog==3.0.0

# ─── PDF Generation ───────────────────────────────────────────────────────────
weasyprint==68.1

# ─── Image Handling ───────────────────────────────────────────────────────────
Pillow==10.4.0

# ─── Filtering ────────────────────────────────────────────────────────────────
django-filter==24.3

# ─── Rate Limiting ────────────────────────────────────────────────────────────
django-ratelimit==4.1.0

# ─── Production Server ────────────────────────────────────────────────────────
gunicorn==23.0.0
